==> 2. Get Bechdel test data.ipynb <==
import urllib
import json
import pandas as pd
import html
print('Done')

bechdel_url = 'http://bechdeltest.com/api/v1/getAllMovieIds'
bechdel_page = urllib.request.urlopen(bechdel_url).read().decode('utf-8')
#print(bechdel_page)
movie_ids = json.loads(bechdel_page)
bechdel_ids = [movie[u'id'] for movie in movie_ids]
imdb_ids = [movie[u'imdbid'] for movie in movie_ids]
print(u"There are {0:,.0f} movies in the Bechdel test corpus.".format(len(bechdel_ids)))
print(u"{0:,.0f} of them have IMDb IDs.".format(len(imdb_ids)))

bechdel_url = 'http://bechdeltest.com/api/v1/getMovieByImdbId?imdbid=0235712'
bechdel_page = urllib.request.urlopen(bechdel_url).read().decode('utf-8')
bechdel_page = html.unescape(bechdel_page)
x = json.loads(bechdel_page)
x
#imdb_ids

bechdel_url = 'http://bechdeltest.com/api/v1/getAllMovieIds'
bechdel_page = urllib.request.urlopen(bechdel_url).read().decode('utf-8')
#print(bechdel_page)
movie_ids = json.loads(bechdel_page)
bechdel_ids = [movie[u'id'] for movie in movie_ids]
imdb_ids = [movie[u'imdbid'] for movie in movie_ids]
print(u"There are {0:,.0f} movies in the Bechdel test corpus.".format(len(bechdel_ids)))
print(u"{0:,.0f} of them have IMDb IDs.".format(len(imdb_ids)))

ratings = dict()
exceptions = list()


totalmovies = len(imdb_ids)
#i = 0
#for num,imdb_id in enumerate(imdb_ids):
#for i in range(1,5):
for i in range(1,totalmovies):
    imdb_id = imdb_ids[i]
    try:
        bechdel_url = 'http://bechdeltest.com/api/v1/getMovieByImdbId?imdbid={0}'.format(imdb_id)
#        print(bechdel_url)
        bechdel_info = urllib.request.urlopen(bechdel_url).info()
        bechdel_page = urllib.request.urlopen(bechdel_url).read().decode('ISO-8859-1')
        bechdel_page = html.unescape(bechdel_page)
#        print(bechdel_info)
        ratings[imdb_id] = json.loads(bechdel_page)
    except:
        exceptions.append(imdb_id)
        pass
#    i = i + 1
    if ((i % 500 == 0) | (i == totalmovies)):
        print('Processing movie number {0:,.0f} of {1:,.0f}...'.format(i,totalmovies))
print('Movies processed!')

del ratings['']

print('Creating dataframe...')
ratings_df = pd.DataFrame.from_dict(ratings,orient='index')
ratings_df.index.name = 'imdb_id'
ratings_df = ratings_df.drop(['imdbid'],axis=1)
ratings_df.columns = ['bechdel_title','bechdel_submitter_id','bechdel_submitted_date','bechdel_dubious','bechdel_rating','bechdel_year','bechdel_site_id','bechdel_visible']
#ratings_df

print('\n')
print('{0:,.0f} movies written. Exceptions below.'.format(len(ratings_df)))
for i in range(0, len(exceptions)):
    print(i, exceptions[i])

print('Writing CSV...')
ratings_df.to_csv('bechdel.csv',encoding='utf-8')
print('Done')

##ratings_bk = ratings
#ratings = ratings_bk
for i in range(16,len(exceptions)):
    print(i, exceptions[i],'http://www.imdb.com/title/tt{0:}'.format(exceptions[i]))
