==> 3. Get data from OMDb.ipynb <==
import urllib.request
import json
import pandas as pd
import numpy as np
from pprint import pprint
bechdel_data = pd.read_csv('bechdel.csv',encoding='utf-8')
bechdel_data = bechdel_data.set_index('imdb_id')
len(bechdel_data)
#bechdel_data.sample(3)        
#testurl = 'http://www.omdbapi.com/?i=tt0848228&apikey=949d45c1&plot=full'  # The Avengers (2012)
#page = urllib.request.urlopen(testurl).read().decode('utf-8')
#x = json.loads(page,encoding='utf-8')
#pprint(x)

omdb_data = dict()
exceptions = list()
id_list = bechdel_data.index.tolist()
#id_list
totalmovies = len(id_list)
print(totalmovies)
i = 0
for theid in id_list:
    i = i + 1
    id_string = 'tt'    
    needed_zeroes = 7 - len(str(theid))
    for j in range(0,needed_zeroes):
        id_string = id_string + '0'
    id_string = id_string + str(theid)
    omdburl = 'http://www.omdbapi.com/?i={0}&apikey=949d45c1'.format(id_string)
    try:
        omdbpage = urllib.request.urlopen(omdburl).read().decode('utf-8')
        omdb_data[theid] = json.loads(omdbpage,encoding='utf-8')
    except:
        exceptions.append(theid)
        pass
    if ((i % 500 == 0) | (i == totalmovies)):
        print('Processing movie {0:,.0f} of {1:,.0f}...'.format(i,totalmovies))
print(exceptions)
omdb_df = pd.DataFrame.from_dict(omdb_data,orient='index')
omdb_df.index.name = 'imdb_id'
#omdb_df.sample(3)
#omdb_df.to_csv('omdb.csv',encoding='utf-8')

# Drop non-movie-related columns
omdb_df = omdb_df.drop(['Season','seriesID','Episode','totalSeasons'],axis=1)

# Rename columns to what we want them to be.
omdb_df=omdb_df.rename(columns = {'Actors':'actors','Writer':'writers','Response':'response','Language':'languages'})
omdb_df=omdb_df.rename(columns = {'Title':'imdbtitle','Rated':'mpaa_rating','imdbRating':'userscore','Type':'type'})
omdb_df=omdb_df.rename(columns = {'Poster':'posterlink','BoxOffice':'imdb_boxoffice','Country':'countries'})
omdb_df=omdb_df.rename(columns = {'Ratings':'all_ratings_list','Genre':'genres','Runtime':'runtime','Metascore':'metascore'})
omdb_df=omdb_df.rename(columns = {'Awards':'awards','Plot':'imdb_plot','DVD':'dvd_release_date','Director':'director'})
omdb_df=omdb_df.rename(columns = {'Website':'website','Year':'imdbyear','Released':'imdb_release_date','imdbID':'imdb_id_str'})
omdb_df=omdb_df.rename(columns = {'Production':'studio','imdbVotes':'nVotes'})

# Drop non-movies
omdb_df = omdb_df[omdb_df['type'] == 'movie']


# Fix runtimes to be integers
omdb_df['runtime'] = omdb_df['runtime'].str.slice(stop=-4).replace('',np.nan)
omdb_df['runtime'] = omdb_df['runtime'].apply(float)
omdb_df['runtime'] = omdb_df['runtime'].apply(lambda x: '{:.0f}'.format(x))

# Fix release dates to be datetime, formatted as yyyy-mm-dd
omdb_df['imdb_release_date'] = pd.to_datetime(omdb_df['imdb_release_date'],format="%d %b %Y", errors='coerce')

# Convert Release dateimte into new columns for year, month, and week
#omdb_df['imdb_release_year'] = omdb_df['imdb_release_date'].dt.strftime('%Y')
#omdb_df['imdb_release_month'] = omdb_df['imdb_release_date'].dt.strftime('%B')
#omdb_df['imdb_release_weekday'] = omdb_df['imdb_release_date'].dt.strftime('%A')

#omdb_df['imdb_release_week'] = omdb_df['imdb_release_date'].apply(lambda x:x.week)

# Take the imdbVotes formatted as string containing "N/A" and comma-delimited thousands, convert to float
omdb_df['nVotes'] = omdb_df['nVotes'].dropna().replace('N/A',np.nan).dropna().apply(lambda x:float(x.replace(',','')))

#imdb_df[['imdb_id','imdbtitle','mpaa_rating','type','imdb_release_date','imdb_release_year','imdb_release_month','imdb_release_week','imdb_release_weekday']]
#omdb_df['nVotes'] = omdb_df['nVotes'].apply(lambda x: '{:,.0f}'.format(x))

# Take the Metascore formatted as string containing "N/A", convert to float
omdb_df['metascore'] = omdb_df['metascore'].dropna().replace('N/A',np.nan).dropna().apply(float)

# Also divide by 10 to make effect sizes more comparable
#omdb_df['metascore'] = omdb_df['metascore'] / 10

# Take the imdbRating formatted as string containing "N/A", convert to float
omdb_df['userscore'] = omdb_df['userscore'].dropna().replace('N/A',np.nan).dropna().apply(float)

# Create indicator variables for English and USA movies (English is at least one of its languages etc.) (0 = no, 1 =yes)
omdb_df['english'] = (omdb_df['languages'].dropna().str.contains('nglish')).astype(int)
omdb_df['usa'] = (omdb_df['countries'].dropna().str.contains('USA')).astype(int)

#omdb_df.sample(3)
omdb_df.to_csv('omdb.csv',encoding='utf-8')

print('Done')
#omdb_df.sample(10)



import numpy as np
#import pandas as pd
import time
omdb_df = pd.read_csv('omdb.csv',encoding='utf-8') 









omdb_df.sample(3)
#omdb_df.loc[3].transpose()
omdb_df.to_csv('omdb.csv',encoding='utf-8')
print('Done')
