==> 4. Get Box Office Mojo data.ipynb <==
import pandas as pd
import numpy as np

boxofficemojo_df = pd.read_csv('boxofficemojo2000s_in.csv')
boxofficemojo_df = boxofficemojo_df.set_index(['key'])

boxofficemojo_df = boxofficemojo_df.rename(columns = {'opening_year':'year','opening_month':'month','opening_day':'day'})
boxofficemojo_df[['originalorder','totalgross','totaltheaters','openinggross','openingtheaters','year','month','day']] = boxofficemojo_df[['originalorder','totalgross','totaltheaters','openinggross','openingtheaters','year','month','day']].replace([u'',u'nan'],[np.nan,np.nan])

boxofficemojo_df['totalgross'] = boxofficemojo_df['totalgross'].dropna().str.replace('$','').str.replace(',','')
boxofficemojo_df['totalgross'] = pd.to_numeric(boxofficemojo_df['totalgross'])
boxofficemojo_df['openinggross'] = boxofficemojo_df['openinggross'].dropna().str.replace('$','').str.replace(',','')
boxofficemojo_df['openinggross'] = pd.to_numeric(boxofficemojo_df['openinggross'])

boxofficemojo_df['totaltheaters'] = boxofficemojo_df['totaltheaters'].dropna().str.replace(',','')
boxofficemojo_df['totaltheaters'] = pd.to_numeric(boxofficemojo_df['totaltheaters'])
boxofficemojo_df['openingtheaters'] = boxofficemojo_df['openingtheaters'].dropna().str.replace(',','')
boxofficemojo_df['openingtheaters'] = pd.to_numeric(boxofficemojo_df['openingtheaters'])

boxofficemojo_df['release_date'] = pd.to_datetime(boxofficemojo_df[['year', 'month', 'day']])#,errors=coerce)
#print(pd.to_datetime(boxofficemojo_df[['year', 'month', 'day']]),errors=coerce)
#print('done')
#boxofficemojo_df[['key','title','year','month','day']]

boxofficemojo_df.sort_index()

#boxofficemojo_df.groupby(['day']).size()
#boxofficemojo_df[boxofficemojo_df['key']=='zombied11dawn']
#boxofficemojo_df.loc[boxofficemojo_df['year']==15]
boxofficemojo_df = boxofficemojo_df.drop(['year','month','day','originalorder'],axis=1)
boxofficemojo_df.to_csv('boxofficemojo2000s.csv')
print('Done')
