==> 6. Get variables.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime
print('Done!')

movies = pd.read_csv('movies.csv',index_col='imdb_id')

movies['imdb_boxoffice'] = movies['imdb_boxoffice'].str.replace('$','').str.replace(',','')
movies['imdb_boxoffice'] = pd.to_numeric(movies['imdb_boxoffice'])

#print('Movies with IMDb box office data: ',movies['imdb_boxoffice'][(~pd.isnull(movies['imdb_boxoffice']))].size)
#print('Movies with boxofficemojo box office data: ',movies['totalgross'][(~pd.isnull(movies['totalgross']))].size)
#print('Movies with the-numbers box office data: ',movies['revenue'][(~pd.isnull(movies['revenue']))].size)
#print('\n')
#print('Movies with all three:',len(movies[['bechdel_title','date','imdb_boxoffice','totalgross','revenue']][(~pd.isnull(movies['imdb_boxoffice'])) & (~pd.isnull(movies['totalgross'])) & (~pd.isnull(movies['revenue']))]))
#print('Movies with any one or more of the three:',len(movies[['bechdel_title','date','imdb_boxoffice','totalgross','revenue']][(~pd.isnull(movies['imdb_boxoffice'])) | (~pd.isnull(movies['totalgross'])) | (~pd.isnull(movies['revenue']))]))

# Let's use the numbers where available, since that's what Keegan used.
# Order of trustworthiness: the-numbers, boxofficemojo, IMDb
movies['boxoffice'] = movies['revenue_thenumbers']
movies['boxoffice'] = np.where(pd.isnull(movies['boxoffice']),movies['totalgross'],movies['boxoffice'])
movies['boxoffice'] = np.where(pd.isnull(movies['boxoffice']),movies['imdb_boxoffice'],movies['boxoffice'])

print('Movies with box office receipt data: {0:,.0f}'.format(len(movies[~pd.isnull(movies['boxoffice'])])))

movies['budget'] = movies['budget_thenumbers']
print('Movies with budget/profit data: {0:,.0f}'.format(len(movies[~pd.isnull(movies['budget'])])))


movies['profit'] = movies['boxoffice'] - movies['budget']
#movies[['bechdel_title','year','boxoffice','budget','profit']][~pd.isnull(movies['profit'])].style.format({'bechdel_title': '{:.40}', 'year': '{:.0f}', 'boxoffice': '${:,.0f}', 'budget': '${:,.0f}', 'profit': '${:,.0f}' })
movies.sample(1).transpose()

print('Movies with IMDb release dates: ',movies['imdb_release_date'][(~pd.isnull(movies['imdb_release_date']))].size)
print('Movies with boxofficemojo release dates: ',movies['release_date'][(~pd.isnull(movies['release_date']))].size)
print('Movies with the-numbers release dates: ',movies['releasedate_thenumbers'][(~pd.isnull(movies['releasedate_thenumbers']))].size)

print('Movies with all three:',len(movies[['imdb_release_date','release_date','releasedate_thenumbers']][(~pd.isnull(movies['imdb_release_date'])) & (~pd.isnull(movies['release_date'])) & (~pd.isnull(movies['releasedate_thenumbers']))]))
print('Movies with all three:',len(movies[['imdb_release_date','release_date','releasedate_thenumbers']][(~pd.isnull(movies['imdb_release_date'])) | (~pd.isnull(movies['release_date'])) | (~pd.isnull(movies['releasedate_thenumbers']))]))

movies['released'] = movies['releasedate_thenumbers']
movies['released'] = np.where(pd.isnull(movies['released']),movies['release_date'],movies['released'])
movies['released'] = np.where(pd.isnull(movies['released']),movies['imdb_release_date'],movies['released'])

movies[['bechdel_title','released']][~pd.isnull(movies['released'])].sample(3)
#movies[['bechdel_title','imdb_release_date','release_date','releasedate']][(~pd.isnull(movies['imdb_release_date'])) & (~pd.isnull(movies['release_date'])) & (~pd.isnull(movies['releasedate']))].sample(30)

cpi = pd.read_csv('inflation/cpi-1913-2017.csv',index_col='Year')
#inflator = dict(cpi.ix[2014,'Annual']/cpi['Annual'])
cpi = cpi.drop(['HALF1','HALF2'],axis=1)
#cpi.size
dec2016 = cpi.get_value(2016,'Dec')

inflator_array = np.empty((cpi.size,3),dtype=object)
inflator_array.shape
i = 0
for j in cpi.index:
    for k in cpi.columns:
        inflator_array[i,0] = j
        inflator_array[i,1] = str(k)
        inflator_array[i,2] = cpi.loc[j][k] / dec2016
        i = i + 1
#for m in range(0,len(inflator)):
#    for n in range(0,3):
#        print(inflator[m,n])
#dec2016 = cpi.get_value(2016,'Dec')
#print(dec2016)
#y
inflator = pd.DataFrame(inflator_array,columns=[['year','month','factor']])
inflator['year'] = pd.to_numeric(inflator['year'])
inflator['month'] = inflator['month'].astype(str)
inflator['factor'] = pd.to_numeric(inflator['factor'])
inflator.dtypes

movies['released'] = pd.to_datetime(movies['released'])
movies['releaseyear'] = pd.to_datetime(movies['released'].dropna().apply(lambda x: x.year))
movies['releasemonth'] = movies['released'].dropna().apply(lambda x: x.strftime('%b')).astype(str)
movies.dtypes
#print(movies.shape)

# merge to preserve index from left table, cargo-culted from https://stackoverflow.com/questions/11976503/how-to-keep-index-when-using-pandas-merge
#joined_df_con_llaves = joined_df.reset_index().merge(llaves_df, how='left', on=['imdbtitle','imdbyear']).set_index('imdb_id')
movies_inflated = movies.reset_index().merge(inflator,how='left',left_on=['releaseyear','releasemonth'],right_on=['year','month']).set_index('imdb_id')

#movies_inflated[['bechdel_title','released','factor']]
#print(movies_inflated.columns)

movies_inflated['boxoffice_adj'] = movies_inflated['boxoffice'] / movies_inflated['factor']
movies_inflated['budget_adj'] = movies_inflated['budget'] / movies_inflated['factor']
movies_inflated['openinggross_adj'] = movies_inflated['openinggross'] / movies_inflated['factor']
movies_inflated['profit_adj'] = movies_inflated['boxoffice_adj'] - movies_inflated['budget_adj']
movies_inflated['roi_adj'] = movies_inflated['profit_adj'] / movies_inflated['budget_adj']

#movies_inflated[['bechdel_title','released','boxoffice','boxoffice_adj','budget','budget_adj','profit','profit_adj']].sample(3)#.style.format({'bechdel_title': '{:.40}', 'released': str, 'boxoffice': '${:,.0f}', 'boxoffice_adj': '${:,.0f}','budget': '${:,.0f}', 'budget_adj': '${:,.0f},})


movies_inflated['title'] = movies_inflated['bechdel_title']

movies_inflated = movies_inflated.drop('imdb_link.1', axis=1)
movies_inflated = movies_inflated.drop('releaseyear', axis=1)
movies_inflated = movies_inflated.drop('releasemonth', axis=1)
movies_inflated = movies_inflated.drop('year', axis=1)
movies_inflated = movies_inflated.drop('month', axis=1)

#print(movies_inflated.shape)
#movies_inflated.sample(3)

print('old columns: {0:.0f}'.format(len(movies_inflated.columns)))
#print(movies_inflated.columns)

#print(movies_inflated.columns)

new_column_order_list = list()
# most important
new_column_order_list += ['llave', 'title', 'released', 'bechdel_rating', 'bechdel_dubious']
new_column_order_list += ['boxoffice_adj', 'budget_adj', 'profit_adj', 'roi_adj', 'openinggross_adj', 'totaltheaters', 'openingtheaters']
new_column_order_list += ['boxoffice', 'budget', 'profit', 'openinggross']
# less important
new_column_order_list += ['mpaa_rating','studio', 'nVotes', 'userscore', 'metascore', 'genres', 'runtime', 'english', 'usa', 'imdb_plot']
new_column_order_list += ['awards', 'director', 'writers', 'actors', 'posterlink', 'website', 'dvd_release_date', 'languages', 'countries', 'all_ratings_list']
# Bechdel-specific
new_column_order_list += ['bechdel_title', 'bechdel_submitter_id', 'bechdel_submitted_date', 'bechdel_year', 'bechdel_site_id', 'bechdel_visible']
# IMDb-specific
new_column_order_list += ['imdbtitle', 'imdbyear', 'imdb_release_date', 'imdb_boxoffice', 'imdb_link']
# Boxofficemojo-specific
new_column_order_list += ['title_boxofficemojo', 'studio_boxofficemojo', 'totalgross', 'boxofficemojo_url', 'release_date']
# the-numbers-specific
new_column_order_list += ['releasedate_thenumbers', 'title_thenumbers', 'genre_thenumbers', 'budget_thenumbers', 'revenue_thenumbers', 'year_thenumbers',  'has_trailer_thenumbers', 'the_numbers_url','factor']

movies_inflated = movies_inflated[new_column_order_list]

print('new columns: {0:.0f}'.format(len(movies_inflated.columns)))
#print(movies_inflated.columns)
movies_inflated.to_csv('movies_for_analysis.csv',encoding='utf-8')
print('Done')
